requirements.txt
# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5

# HTTP Client
httpx>=0.24.0

# Markdown Processing
markdown-it-py[linkify,plugins]

# Image Processing & Web Scraping
Pillow>=9.0.0

# API Integration
anthropic>=0.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
ebooklib>=0.17.1
html2text>=2020.1.16

# Environment & Utils
python-dotenv>=0.19.0
requests>=2.26.0

# Development & Testing
black>=22.3.0
pytest>=7.0.0
pytest-cov>=3.0.0 